+++ src/EventEmitter.cpp
@@ -168,6 +168,12 @@ namespace be {
         return jscotr ;
     }
 
+    void EventEmitter::emitSyncFree(const char * eventName, std::initializer_list<JSValue> args) {
+        emitSync(eventName, args) ;
+        for(auto arg : args) {
+            JS_FreeValue(ctx, arg) ;
+        }
+    }
     void EventEmitter::emitSync(const char * eventName, std::initializer_list<JSValue> args) {
         JSValue name = JS_NewString(ctx, eventName) ;
         emitSync(name, args) ;
@@ -221,11 +227,14 @@ namespace be {
         JSEngine::fromJSContext(ctx)->addLoopFunction((EngineLoopFunction)nativeEventLoop, this, true) ;
     }
 
-    void EventEmitter::emitNativeEvent(void * param) {
+    // 注意，该函数可能返回 false ，
+    // 消息没有实际进入到队列中，因为队列满了或队列尚未创建
+    // 如果消息中有由接收方负责释放的指针，此时需要由发送方负责释放回收
+    bool EventEmitter::emitNativeEvent(void * param) {
         if(!nevent_queue) {
-            return ;
+            return false ;
         }
-        xQueueSend(nevent_queue, param, 0) ;
+        return xQueueSend(nevent_queue, param, 0) == pdTRUE ;
     }
     
     void EventEmitter::nativeEventLoop(JSContext * ctx, EventEmitter * ee) {