+++ src/EventEmitter.hpp
@@ -31,11 +31,12 @@ namespace be {
         EventEmitter(JSContext * ctx, JSValue _jsobj=JS_NULL) ;
         ~EventEmitter() ;
         
+        void emitSyncFree(const char * eventName, std::initializer_list<JSValue> args) ;
         void emitSync(const char * eventName, std::initializer_list<JSValue> args) ;
         void emitSync(const JSValue & eventName, std::initializer_list<JSValue> args) ;
 
 #ifdef ESP_PLATFORM
-        void emitNativeEvent(void * param) ;
+        bool emitNativeEvent(void * param) ;
 #endif
 
         static JSValue constructor(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;